requirements.txt
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
scikit_learn==1.1.2
streamlit==1.12.0
